Move DbFunctions from mathesar to db namespace
