+++ mathesar/api/viewsets/functions.py
@@ -1,6 +1,6 @@
 from rest_framework import viewsets
 from rest_framework.response import Response
-from mathesar.database.functions.base import supported_db_functions
+from db.functions.base import supported_db_functions
 from mathesar.api.serializers.functions import DbFunctionSerializer
 
 class DbFunctionViewSet(viewsets.ViewSet):
