+++ mathesar/database/types.py
@@ -5,9 +5,6 @@ Mathesar data types are shown in the UI.
 from enum import Enum
 
 from db.types.base import PostgresType, MathesarCustomType, get_available_types, get_qualified_name, get_db_type_name
-from db.filters.base import Predicate, relies_on_comparability
-
-from typing import Type
 
 
 class MathesarTypeIdentifier(Enum):
@@ -23,25 +20,6 @@ class MathesarTypeIdentifier(Enum):
     OTHER = 'other'
     # These are types that we don't know anything about.
     CUSTOM = 'custom'
-
-
-comparable_mathesar_types = {
-    MathesarTypeIdentifier.DATETIME,
-    MathesarTypeIdentifier.DURATION,
-    MathesarTypeIdentifier.MONEY,
-    MathesarTypeIdentifier.NUMBER,
-}
-
-
-def _is_ma_type_comparable(ma_type: MathesarTypeIdentifier) -> bool:
-    return ma_type in comparable_mathesar_types
-
-
-def is_ma_type_supported_by_predicate(ma_type: MathesarTypeIdentifier, predicate: Type[Predicate]):
-    if relies_on_comparability(predicate):
-        return _is_ma_type_comparable(ma_type)
-    else:
-        return True
 
 
 def _get_mapped_types(type_map):
